2/9_laravel_テーブル関連付け

+++ php+laravel/example-app/app/Http/Controllers/Tweet/DeleteControler.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Tweet;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Tweet;
+use App\Services\TweetService;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteControler extends Controller
 {
@@ -14,9 +16,12 @@ class DeleteControler extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request,TweetService $tweetService)
     {
         $tweetid = (int) $request->route("tweetId");
+        if(!$tweetService->checkOwnTweet($request->user()->id,$tweetid)) {
+            throw new AccessDeniedHttpException();
+        }
         $tweet = Tweet::where("id",$tweetid)->firstOrFail();
         $tweet->delete();
         return redirect()
